restructuring  of post tests file  for better usage
