+++ test/post.test.js
@@ -36,7 +36,7 @@ describe("Posts API", () => {
      * providing correct post id
      */
     it("should get post by id", () => {
-      const id = "5f46c9c4b5512991dbd36352";
+      const id = "5f4663e66450ec01f0adeed2";
       chai
         .request(server)
         .get("/api/posts/" + id)
@@ -55,7 +55,7 @@ describe("Posts API", () => {
      * providing wrong id
      */
     it("shouldn't get any post", () => {
-      const id2 = "5f46c9c4b5512991dbd3635";
+      const id2 = "5f4663e66450ec01f0adeed";
       chai
         .request(server)
         .get("/api/posts/" + id2)
@@ -66,268 +66,5 @@ describe("Posts API", () => {
           
         });
     });
-  });
-
-  /*
-   * POST one post
-   */
-  describe("POST api/posts", () => {
-    /*
-     * providing correct routes and  correct info
-     */
-    it("should post one post", () => {
-      let post = {
-        title: "new title",
-        content: "new content for testing",
-        image: "image location",
-      };
-
-      const adminToken =
-        "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImhlbGxvMkBnbWFpbC5jb20iLCJpc2FkbWluIjp0cnVlLCJpYXQiOjE1OTg1MjAyNDcsImV4cCI6MTU5ODYxMzg0N30.jOTouEUaWQ6rs_bKgMM6BgX2pc-J9Z0DPzGz0eaR-7w";
-      chai
-        .request(server)
-        .post("/api/posts/")
-        .set({ Authorization: adminToken })
-        .send(post)
-        .end((err, response) => {
-          response.should.have.status(201);
-          response.body.should.be.a("object");
-          response.body.should.have
-            .property("message")
-            .eq("post created successfuly");
-          response.body.should.have.property("data");
-
-          
-        });
-    });
-
-    /*
-     * providing wrong tokken
-     */
-    it("should not be able to post any post", (done) => {
-      const post = {
-        title: "new times roman",
-        content: "new content for testing",
-        img: "image location",
-      };
-      const token =
-        "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImhlbGxvMkBnbWFpbC5jb20iLCJpc2FkbWluIjp0cnVlLCJpYXQiOjE1OTg1MjE0MTksImV4cCI6MTU5ODYxNTAxOX0.wbMcRNcUuZqUYp-kgFvPE_ZRJWY41gjxBq1knmWUj_A";
-      chai
-        .request(server)
-        .post("/api/posts/")
-        .set({ Authorization: token })
-        .send(post)
-        .end((err, response) => {
-          response.should.have.status(401);
-          response.body.should.be.a("object");
-          response.body.should.have.property("error").eq("invalid token");
-          done();
-        });
-    });
-
-    /*
-     * providing unauthorized User
-     */
-    it("should not be able to post", (done) => {
-      const post = {
-        title: "new title for testing",
-        content: "new content for testing",
-        img: "image location",
-      };
-      const reguralToken =
-        "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImhlbGxvQGdtYWlsLmNvbSIsImlzYWRtaW4iOmZhbHNlLCJpYXQiOjE1OTg0NzMwNTUsImV4cCI6MTU5ODQ5NDY1NX0.rfoMVxk3an2rZ0p9_oxYRRfXpYYkjwxpY6IAE3JMElI";
-      chai
-        .request(server)
-        .post("/api/posts/")
-        .set({ Authorization: reguralToken })
-        .send(post)
-        .end((err, response) => {
-          response.should.have.status(403);
-          response.body.should.be.a("object");
-          response.body.should.have
-            .property("error")
-            .eq("Only admins can access this");
-          done();
-        });
-    });
-
-    /*
-     * providing empty fields
-     */
-    it("should not be able to post", (done) => {
-      const post = {
-        title: "new title for testing",
-        content: "new content for testing",
-        content: "",
-      };
-      const adminToken =
-        "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImhlbGxvMkBnbWFpbC5jb20iLCJpc2FkbWluIjp0cnVlLCJpYXQiOjE1OTg1MjAyNDcsImV4cCI6MTU5ODYxMzg0N30.jOTouEUaWQ6rs_bKgMM6BgX2pc-J9Z0DPzGz0eaR-7w";
-      chai
-        .request(server)
-        .post("/api/posts/")
-        .set({ Authorization: adminToken })
-        .send(post)
-        .end((err, response) => {
-          response.should.have.status(400);
-          response.body.should.be.a("object");
-          response.body.should.have
-            .property("message")
-            .eq("please fill in all the fields");
-          done();
-        });
-    });
-
-    /*
-     * providing missing fields
-     */
-    it("should not be able to post", (done) => {
-      const post = {
-        title: "new title for testing",
-      };
-      const adminToken =
-        "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImhlbGxvMkBnbWFpbC5jb20iLCJpc2FkbWluIjp0cnVlLCJpYXQiOjE1OTg1MjAyNDcsImV4cCI6MTU5ODYxMzg0N30.jOTouEUaWQ6rs_bKgMM6BgX2pc-J9Z0DPzGz0eaR-7w";
-      chai
-        .request(server)
-        .post("/api/posts/")
-        .set({ Authorization: adminToken })
-        .send(post)
-        .end((err, response) => {
-          response.should.have.status(400);
-          response.body.should.be.a("object");
-          response.body.should.have
-            .property("message")
-            .eq("please fill in all the fields");
-          done();
-        });
-    });
-  });
-
-  /*
-   * UPDATE post
-   */
-  describe("Update api/posts/id", () => {
-    /*
-     * providing correct routes and  correct info
-     */
-    it("should update a post", (done) => {
-      let post = {
-        title: "new title updated",
-        content: "new content for testing updated ",
-        image: "image location updated",
-      };
-      let id = "5f46c9c4b5512991dbd36352";
-      const adminToken =
-        "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImhlbGxvMkBnbWFpbC5jb20iLCJpc2FkbWluIjp0cnVlLCJpYXQiOjE1OTg1MjAyNDcsImV4cCI6MTU5ODYxMzg0N30.jOTouEUaWQ6rs_bKgMM6BgX2pc-J9Z0DPzGz0eaR-7w";
-      chai
-        .request(server)
-        .patch("/api/posts/" + id)
-        .set({ Authorization: adminToken })
-        .send(post)
-        .end((err, response) => {
-          response.should.have.status(200);
-          response.body.should.be.a("object");
-          response.body.should.have
-            .property("message")
-            .eq("post updated succeddfuly");
-          response.body.should.have.property("data");
-
-          done();
-        });
-    });
-
-    /*
-     * providing unauthorized User
-     */
-    it("should not be able to update ", (done) => {
-      const post = {
-        title: "new title for testing up",
-        content: "new content for testing up",
-        image: "image location",
-      };
-      const id = "5f46e31d776c67d88e661817";
-      const reguralToken =
-        "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImhlbGxvQGdtYWlsLmNvbSIsImlzYWRtaW4iOmZhbHNlLCJpYXQiOjE1OTg1MjY4MjAsImV4cCI6MTU5ODYyMDQyMH0.O3thjzSRZ9lYm8XLMiekilkZCqoaJji5Pl5yvhi6rNY";
-      chai
-        .request(server)
-        .patch("/api/posts/" + id)
-        .set({ Authorization: reguralToken })
-        .send(post)
-        .end((err, response) => {
-          response.should.have.status(403);
-          response.body.should.be.a("object");
-          response.body.should.have
-            .property("error")
-            .eq("Only admins can access this");
-          done();
-        });
-    });
-
-    /*
-     * providing wrong post id
-     */
-    it("should not be able to update", (done) => {
-      const post = {
-        title: "new title for testing update",
-        content: "new content for testing update",
-        image: "image location",
-      };
-      const id = "5f46e31d776c67d88e66181";
-      const adminToken =
-        "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImhlbGxvMkBnbWFpbC5jb20iLCJpc2FkbWluIjp0cnVlLCJpYXQiOjE1OTg1MjAyNDcsImV4cCI6MTU5ODYxMzg0N30.jOTouEUaWQ6rs_bKgMM6BgX2pc-J9Z0DPzGz0eaR-7w";
-      chai
-        .request(server)
-        .patch("/api/posts/" + id)
-        .set({ Authorization: adminToken })
-        .send(post)
-        .end((err, response) => {
-          response.should.have.status(404);
-          response.body.should.be.a("object");
-          response.body.should.have.property("error").eq("Post doesn't exist!");
-          done();
-        });
-    });
-  });
-
-  /*
-   * DELETE Post
-   */
-  describe("DELETE api/posts/id", () => {
-    /**
-     * Provide collect id
-     */
-    it("should Delete a post", (done) => {
-      let id = "5f46e487950eeddbb9695fbc";
-      const adminToken =
-        "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImhlbGxvMkBnbWFpbC5jb20iLCJpc2FkbWluIjp0cnVlLCJpYXQiOjE1OTg1MjAyNDcsImV4cCI6MTU5ODYxMzg0N30.jOTouEUaWQ6rs_bKgMM6BgX2pc-J9Z0DPzGz0eaR-7w";
-      chai
-        .request(server)
-        .delete("/api/posts/" + id)
-        .set({ Authorization: adminToken })
-        .end((err, response) => {
-          response.should.have.status(200);
-          response.body.should.have
-            .property("message")
-            .eq("post deleted successfuly");
-          done();
-        });
-    });
-
-    /**
-     * Provide incollect id
-     */
-    it("should Delete a post", (done) => {
-      let id = "5f46e487950eeddbb9695fb";
-      const adminToken =
-        "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImhlbGxvMkBnbWFpbC5jb20iLCJpc2FkbWluIjp0cnVlLCJpYXQiOjE1OTg1MjAyNDcsImV4cCI6MTU5ODYxMzg0N30.jOTouEUaWQ6rs_bKgMM6BgX2pc-J9Z0DPzGz0eaR-7w";
-      chai
-        .request(server)
-        .delete("/api/posts/" + id)
-        .set({ Authorization: adminToken })
-        .end((err, response) => {
-          response.should.have.status(404);
-          response.body.should.have.property("error").eq("Post doesn't exist!");
-          done();
-        });
-    });
-  });
+  });  
 });
